+++ packages/waggle/src/libp2p/websocketOverTor.test.ts
@@ -86,6 +86,12 @@ describe('websocketOverTor', () => {
     ['array', Array]
   ])('connects successfully with CA passed as %s', async (_name: string, caType: (ca: string) => any) => {
     const pems = await createCertificatesTestHelper(`${service1.onionAddress}`, `${service2.onionAddress}`)
+    // In case test fails on CI, we will be able to conduct test against failing credentials.
+    console.log(`serVert ${pems.servCert}`)
+    console.log(`servKey ${pems.servKey}`)
+    console.log(`ca ${pems.ca}`)
+    console.log(`userCert ${pems.userCert}`)
+    console.log(`userKey ${pems.userKey}`)
 
     const prepareListenerArg = {
       handler: (x) => x,
@@ -129,7 +135,7 @@ describe('websocketOverTor', () => {
         agent,
         cert: pems.userCert,
         key: pems.userKey,
-        ca: [pems.ca]
+        ca: caType(pems.ca)
       },
       localAddr: createLibp2pAddress(service2.onionAddress, port2, peerId2, wsType),
       serverOpts: {},
@@ -149,9 +155,23 @@ describe('websocketOverTor', () => {
     const onConnection = jest.fn()
     listener.on('connection', onConnection)
 
-    await ws2.dial(multiAddress, {
-      signal: signal
-    })
+    let retryCount = 0
+
+    const tryDial = async () => {
+      try {
+        await ws2.dial(multiAddress, {
+          signal: signal
+        })
+      } catch (e) {
+        console.log(`catched Error ${e.message as string}, retryCount is ${retryCount}`)
+        if (retryCount < 2) {
+          retryCount++
+          await tryDial()
+        }
+      }
+    }
+
+    await tryDial()
 
     expect(onConnection).toBeCalled()
     expect(onConnection.mock.calls[0][0].remoteAddr).toEqual(remoteAddress)